refactor(reply): relocate reply policy implementation
